+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/database/model/KnownServerUrlEntity.kt
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,18 +14,14 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.internal.raw
+package org.matrix.android.sdk.internal.database.model
 
-import io.realm.annotations.RealmModule
-import org.matrix.android.sdk.internal.database.model.KnownServerUrlEntity
-import org.matrix.android.sdk.internal.database.model.RawCacheEntity
+import io.realm.RealmObject
+import io.realm.annotations.PrimaryKey
 
-/**
- * Realm module for global classes
- */
-@RealmModule(library = true,
-        classes = [
-            RawCacheEntity::class,
-            KnownServerUrlEntity::class
-        ])
-internal class GlobalRealmModule
+internal open class KnownServerUrlEntity(
+        @PrimaryKey
+        var url: String = ""
+) : RealmObject() {
+    companion object
+}