+++ vector/src/main/java/im/vector/app/core/di/VectorComponent.kt
@@ -59,6 +59,7 @@ import im.vector.app.features.settings.VectorPreferences
 import im.vector.app.features.ui.UiStateRepository
 import org.matrix.android.sdk.api.Matrix
 import org.matrix.android.sdk.api.auth.AuthenticationService
+import org.matrix.android.sdk.api.auth.HomeServerHistoryService
 import org.matrix.android.sdk.api.raw.RawService
 import org.matrix.android.sdk.api.session.Session
 import javax.inject.Singleton
@@ -126,6 +127,8 @@ interface VectorComponent {
     fun authenticationService(): AuthenticationService
 
     fun rawService(): RawService
+
+    fun homeServerHistoryService(): HomeServerHistoryService
 
     fun bugReporter(): BugReporter