+++ vector/src/main/java/im/vector/app/features/login/LoginFragment.kt
@@ -37,6 +37,7 @@ import io.reactivex.Observable
 import io.reactivex.functions.BiFunction
 import io.reactivex.rxkotlin.subscribeBy
 import kotlinx.android.synthetic.main.fragment_login.*
+import kotlinx.android.synthetic.main.fragment_login_signup_signin_selection.*
 import org.matrix.android.sdk.api.failure.Failure
 import org.matrix.android.sdk.api.failure.MatrixError
 import org.matrix.android.sdk.api.failure.isInvalidPassword
@@ -50,7 +51,7 @@ import javax.inject.Inject
  * In signup mode:
  * - the user is asked for login and password
  */
-class LoginFragment @Inject constructor() : AbstractLoginFragment() {
+class LoginFragment @Inject constructor() : AbstractSSOLoginFragment() {
 
     private var passwordShown = false
     private var isSignupMode = false
@@ -83,11 +84,13 @@ class LoginFragment @Inject constructor() : AbstractLoginFragment() {
                 SignMode.SignUp             -> {
                     loginField.setAutofillHints(HintConstants.AUTOFILL_HINT_NEW_USERNAME)
                     passwordField.setAutofillHints(HintConstants.AUTOFILL_HINT_NEW_PASSWORD)
+                    loginSocialLoginButtons.mode = SocialLoginButtonsView.Mode.MODE_SIGN_UP
                 }
                 SignMode.SignIn,
                 SignMode.SignInWithMatrixId -> {
                     loginField.setAutofillHints(HintConstants.AUTOFILL_HINT_USERNAME)
                     passwordField.setAutofillHints(HintConstants.AUTOFILL_HINT_PASSWORD)
+                    loginSocialLoginButtons.mode = SocialLoginButtonsView.Mode.MODE_SIGN_IN
                 }
             }.exhaustive
         }
@@ -169,6 +172,19 @@ class LoginFragment @Inject constructor() : AbstractLoginFragment() {
                 ServerType.Unknown   -> Unit /* Should not happen */
             }
             loginPasswordNotice.isVisible = false
+
+            if (state.loginMode is LoginMode.SsoAndPassword) {
+                loginSocialLoginContainer.isVisible = true
+                loginSocialLoginButtons.ssoIdentityProviders = state.loginMode.ssoIdentityProviders
+                loginSocialLoginButtons.listener = object : SocialLoginButtonsView.InteractionListener {
+                    override fun onProviderSelected(id: String?) {
+                        openInCustomTab(state.getSsoUrl(id))
+                    }
+                }
+            } else {
+                loginSocialLoginContainer.isVisible = false
+                loginSocialLoginButtons.ssoIdentityProviders = null
+            }
         }
     }